[DEV-12054] Merge root layout into `[localeCode]/layout`
To have access to the `localeCode` param value
